% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr.R
\name{read_ogr_array_stream}
\alias{read_ogr_array_stream}
\alias{read_ogr_table}
\alias{read_ogr_df}
\title{Read OGR using the Columnar API}
\usage{
read_ogr_array_stream(dsn)

read_ogr_table(dsn)

read_ogr_df(dsn)
}
\arguments{
\item{dsn}{A GDAL datsource name}
}
\description{
Read OGR using the Columnar API
}
