% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfc86-package.R
\docType{package}
\name{rfc86-package}
\alias{rfc86}
\alias{rfc86-package}
\title{rfc86: Tests GDAL RFC 86 Columnar API}
\description{
Tests GDAL's Columnar API in R.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
